+++ CHANGES.rst
@@ -8,6 +8,8 @@ Changelog
   `expandItems` that is a boolean to decide if the items of the section should 
   be expanded or not (default is True).
   [mamico]
+- Add a profile to limit addables on site root
+  [lucabel]
 
 5.0.9 (2024-04-12)
 ------------------